requirements.txt
telethon>=1.34.0
python-dotenv>=1.0.1
gspread>=6.1.0
google-auth>=2.30.0